Bug Fixes, Documentation Improvements, And More Fixed CustomItems with ItemGlow not glowing when the player died Changed CustomFlags to use long instead of int

+++ UncomplicatedCustomItems/Extensions/CustomFlagsExtensions.cs
@@ -8,25 +8,25 @@ namespace UncomplicatedCustomItems.Extensions
     public static class CustomFlagsExtensions
     {
         /// <summary>
-        /// Checks if the provided <see cref="CustomFlags"/> instance contains the specified custom flag value (represented as an integer).
+        /// Checks if the provided <see cref="CustomFlags"/> instance contains the specified CustomFlag value (represented as a <see langword="long"/>).
         /// This is useful for checking flags that might be defined numerically by other plugins and not present in the base <see cref="CustomFlags"/> enum.
         /// </summary>
         /// <param name="flags">The <see cref="CustomFlags"/> instance to check.</param>
-        /// <param name="customFlagValue">The integer value representing the custom flag to check for.</param>
-        /// <returns><see langword="true"/> if the custom flag represented by <paramref name="customFlagValue"/> is set within the <paramref name="flags"/>; otherwise, <see langword="false"/>.</returns>
-        public static bool HasCustomFlag(this CustomFlags flags, int customFlagValue)
+        /// <param name="customFlagValue">The <see langword="long"/> value representing the CustomFlag to check for.</param>
+        /// <returns><see langword="true"/> if the CustomFlag represented by <paramref name="customFlagValue"/> is set within the <paramref name="flags"/>; otherwise, <see langword="false"/>.</returns>
+        public static bool HasCustomFlag(this CustomFlags flags, long customFlagValue)
         {
-            return ((int)flags & customFlagValue) == customFlagValue;
+            return ((long)flags & customFlagValue) == customFlagValue;
         }
 
         /// <summary>
-        /// Adds the specified custom flag value (represented as an integer) to the <see cref="CustomFlags"/> instance and returns the resulting combined flags.
+        /// Adds the specified CustomFlag value (represented as a <see langword="long"/>) to the <see cref="CustomFlags"/> instance and returns the resulting combined flags.
         /// This allows adding flags defined numerically by other plugins.
         /// </summary>
         /// <param name="flags">The original <see cref="CustomFlags"/> instance.</param>
-        /// <param name="customFlagValue">The integer value representing the custom flag to add.</param>
+        /// <param name="customFlagValue">The <see langword="long"/> value representing the CustomFlag to add.</param>
         /// <returns>A new <see cref="CustomFlags"/> value that includes the original <paramref name="flags"/> plus the flag represented by <paramref name="customFlagValue"/>.</returns>
-        public static CustomFlags AddCustomFlag(this CustomFlags flags, int customFlagValue)
+        public static CustomFlags AddCustomFlag(this CustomFlags flags, long customFlagValue)
         {
             return flags | (CustomFlags)customFlagValue;
         }
